Separate video provider and modal video manager. Provide hooks in modal video manager for calling functions during stages in enableVideo.
